Move drop to presentation to drop module
